Add `update` generator and new migration
To be used as `bin/rails solid_queue:update` to copy new migrations.

The new migration links claimed executions to processes with
`process_name`, and replaces the unique index on processes on
`supervisor_id, name` with just `name`, as we'll want the name to
uniquely identify a process because we'll rely on that to release
claimed executions, independently from the supervisor. Even though it
was really unlikely to have a collision on name because these are set
randomly with `SecureRandom.hex(10)`, in this way we guarantee it.

fix